Commiting changes related to only dcc dependency elimination from the code
