+++ score-client/src/main/java/bio/overture/score/client/command/ViewCommand.java
@@ -34,8 +34,8 @@ import bio.overture.score.client.manifest.ManifestService;
 import bio.overture.score.client.metadata.Entity;
 import bio.overture.score.client.metadata.MetadataService;
 import bio.overture.score.client.slicing.SamFileBuilder;
-import bio.overture.score.client.util.VersionUtils;
 import bio.overture.score.client.view.Viewer;
+import bio.overture.score.core.util.VersionUtils;
 import com.beust.jcommander.Parameter;
 import com.beust.jcommander.Parameters;
 import com.google.common.collect.ImmutableList;
