+++ score-client/src/main/java/bio/overture/score/client/config/ClientBanner.java
@@ -17,8 +17,8 @@
  */
 package bio.overture.score.client.config;
 
-import static bio.overture.score.client.util.Joiners.WHITESPACE;
-import static bio.overture.score.client.util.VersionUtils.getScmInfo;
+import static bio.overture.score.core.util.Joiners.WHITESPACE;
+import static bio.overture.score.core.util.VersionUtils.getScmInfo;
 import static com.fasterxml.jackson.databind.SerializationFeature.FAIL_ON_EMPTY_BEANS;
 import static com.google.common.base.MoreObjects.firstNonNull;
 import static com.google.common.base.Strings.padEnd;
