--- score-client/src/main/java/bio/overture/score/client/ssl/DumbX509TrustManager.java
@@ -1,19 +1,0 @@
-package bio.overture.score.client.ssl;
-
-import java.security.cert.CertificateException;
-import java.security.cert.X509Certificate;
-import javax.net.ssl.X509TrustManager;
-
-public class DumbX509TrustManager implements X509TrustManager {
-  public DumbX509TrustManager() {}
-
-  public void checkClientTrusted(X509Certificate[] chain, String authType)
-      throws CertificateException {}
-
-  public void checkServerTrusted(X509Certificate[] chain, String authType)
-      throws CertificateException {}
-
-  public X509Certificate[] getAcceptedIssuers() {
-    return null;
-  }
-}
