+++ score-test/src/test/java/bio/overture/score/test/meta/MetadataClient.java
@@ -17,14 +17,20 @@
  */
 package bio.overture.score.test.meta;
 
-// import static org.icgc.dcc.common.core.util.stream.Collectors.toImmutableList;
+import static bio.overture.score.core.util.Collectors.toImmutableList;
 
+import bio.overture.score.server.security.ssl.SSLCertificateValidation;
+import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.databind.node.ObjectNode;
+import com.google.common.collect.Lists;
 import java.net.URL;
+import java.util.ArrayList;
+import java.util.List;
 import lombok.Getter;
 import lombok.NonNull;
 import lombok.SneakyThrows;
-// import org.icgc.dcc.common.core.security.SSLCertificateValidation;
+import lombok.val;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
@@ -43,7 +49,7 @@ public class MetadataClient {
   public MetadataClient(
       @Value("${metadata.url}") String serverUrl, @Value("${metadata.ssl.enabled}") boolean ssl) {
     if (!ssl) {
-      //      SSLCertificateValidation.disable();
+      SSLCertificateValidation.disable();
     }
 
     this.serverUrl = serverUrl;
@@ -53,42 +59,41 @@ public class MetadataClient {
     return read("/" + objectId);
   }
 
-  //  public List<Entity> findEntities() {
-  //    return readAll("/");
-  //  }
+  public List<Entity> findEntities() {
+    return readAll("/");
+  }
 
-  //  public List<Entity> findEntitiesByGnosId(@NonNull String gnosId) {
-  //    return readAll("?gnosId=" + gnosId);
-  //  }
+  public List<Entity> findEntitiesByGnosId(@NonNull String gnosId) {
+    return readAll("?gnosId=" + gnosId);
+  }
 
   @SneakyThrows
   private Entity read(@NonNull String path) {
     return MAPPER.readValue(resolveUrl(path), Entity.class);
   }
 
-  //  @SneakyThrows
-  //  private List<Entity> readAll(@NonNull String path) {
-  //    val results = Lists.<Entity>newArrayList();
-  //    boolean last = false;
-  //    int pageNumber = 0;
-  //
-  //    while (!last) {
-  //      val url =
-  //          resolveUrl(path + (path.contains("?") ? "&" : "?") + "size=2000&page=" + pageNumber);
-  //      val result = MAPPER.readValue(url, ObjectNode.class);
-  //      last = result.path("last").asBoolean();
-  //      List<Entity> page =
-  //          MAPPER.convertValue(result.path("content"), new TypeReference<ArrayList<Entity>>()
-  // {});
-  //
-  //      results.addAll(page);
-  //      pageNumber++;
-  //    }
+  @SneakyThrows
+  private List<Entity> readAll(@NonNull String path) {
+    val results = Lists.<Entity>newArrayList();
+    boolean last = false;
+    int pageNumber = 0;
 
-  //    // Remove potential duplicates due to inserts on paging:
-  //    // See https://jira.oicr.on.ca/browse/COL-491
-  //    return results.stream().distinct().collect(toImmutableList());
-  //  }
+    while (!last) {
+      val url =
+          resolveUrl(path + (path.contains("?") ? "&" : "?") + "size=2000&page=" + pageNumber);
+      val result = MAPPER.readValue(url, ObjectNode.class);
+      last = result.path("last").asBoolean();
+      List<Entity> page =
+          MAPPER.convertValue(result.path("content"), new TypeReference<ArrayList<Entity>>() {});
+
+      results.addAll(page);
+      pageNumber++;
+    }
+
+    // Remove potential duplicates due to inserts on paging:
+    // See https://jira.oicr.on.ca/browse/COL-491
+    return results.stream().distinct().collect(toImmutableList());
+  }
 
   @SneakyThrows
   private URL resolveUrl(String path) {
